requirements.txt
crewai
langchain-groq
load_dotenv
crewai_tools
streamlit